back/ericwa-experimental: merge r34071 (float->CGFloat)
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/back/branches/ericwa-experimental@34139 72102866-910b-0410-8b05-ffd578937521

+++ Source/gsc/GSContext.m
@@ -318,12 +318,12 @@ static NSMapTable *gtable;
   [gstate GSSetStrokeColorspace: spaceref];
 }
 
-- (void) GSSetFillColor: (const float *)values
+- (void) GSSetFillColor: (const CGFloat *)values
 {
   [gstate GSSetFillColor: values];
 }
 
-- (void) GSSetStrokeColor: (const float *)values
+- (void) GSSetStrokeColor: (const CGFloat *)values
 {
   [gstate GSSetStrokeColor: values];
 }
